+++ packages/devtools_app/benchmark/test_infra/project_root_directory.dart
@@ -1,6 +1,6 @@
-// Copyright 2023 The Chromium Authors. All rights reserved.
+// Copyright 2023 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
+// found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 
 // Note: this code was copied from Flutter gallery
 // https://github.com/flutter/gallery/blob/main/test_benchmarks/benchmarks/project_root_directory.dart