+++ packages/devtools_app/lib/src/framework/release_notes.dart
@@ -1,6 +1,6 @@
-// Copyright 2021 The Chromium Authors. All rights reserved.
+// Copyright 2021 The Flutter Authors
 // Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
+// found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 
 import 'dart:async';
 import 'dart:convert';
@@ -250,7 +250,7 @@ class ReleaseNotesController extends SidePanelController {
     markdown.value = null;
     toggleVisibility(false);
     if (message != null) {
-      _log.warning('Warning: $message');
+      _log.warning(message);
     }
   }