+++ packages/devtools_app/lib/src/shared/charts/chart_controller.dart
@@ -221,19 +221,17 @@ class ChartController extends DisposableController
       if (lastDT.difference(firstDT).inMinutes >= duration.inMinutes) {
         // Grab the duration in minutes passed in.
         final startOfLastNMinutes =
-            // We need this cast to be able to return null if nothing is found.
-            // ignore: unnecessary_cast
-            timestamps
-            .reversed
-            .firstWhereOrNull((timestamp) {
-              final currentDT = DateTime.fromMillisecondsSinceEpoch(timestamp);
-              final diff = lastDT.difference(currentDT);
-              if (diff.inMinutes >= duration.inMinutes) {
-                return true;
-              }
+        // We need this cast to be able to return null if nothing is found.
+        // ignore: unnecessary_cast
+        timestamps.reversed.firstWhereOrNull((timestamp) {
+          final currentDT = DateTime.fromMillisecondsSinceEpoch(timestamp);
+          final diff = lastDT.difference(currentDT);
+          if (diff.inMinutes >= duration.inMinutes) {
+            return true;
+          }
 
-              return false;
-            });
+          return false;
+        });
 
         final ticksVisible =
             startOfLastNMinutes != null