+++ packages/devtools_app/lib/src/standalone_ui/ide_shared/property_editor/property_editor_panel.dart
@@ -9,28 +9,28 @@ import 'package:devtools_app_shared/utils.dart';
 import 'package:dtd/dtd.dart';
 import 'package:flutter/material.dart';
 
-import '../../service/editor/editor_client.dart';
-import '../../shared/analytics/analytics.dart' as ga;
-import '../../shared/analytics/constants.dart';
-import '../../shared/ui/common_widgets.dart';
-import '../ide_shared/property_editor/property_editor_sidebar.dart';
+import '../../../service/editor/editor_client.dart';
+import '../../../shared/analytics/analytics.dart' as ga;
+import '../../../shared/analytics/constants.dart';
+import '../../../shared/ui/common_widgets.dart';
+import 'property_editor_controller.dart';
+import 'property_editor_view.dart';
 
 /// The side panel for the Property Editor.
-class PropertyEditorSidebarPanel extends StatefulWidget {
-  const PropertyEditorSidebarPanel(this.dtd, {super.key});
+class PropertyEditorPanel extends StatefulWidget {
+  const PropertyEditorPanel(this.dtd, {super.key});
 
   final DartToolingDaemon dtd;
 
   @override
-  State<PropertyEditorSidebarPanel> createState() =>
-      _PropertyEditorSidebarPanelState();
+  State<PropertyEditorPanel> createState() => _PropertyEditorPanelState();
 }
 
-class _PropertyEditorSidebarPanelState
-    extends State<PropertyEditorSidebarPanel> {
-  _PropertyEditorSidebarPanelState();
+class _PropertyEditorPanelState extends State<PropertyEditorPanel> {
+  _PropertyEditorPanelState();
 
   Future<EditorClient>? _editor;
+  PropertyEditorController? _propertyEditorController;
 
   @override
   void initState() {
@@ -38,7 +38,18 @@ class _PropertyEditorSidebarPanelState
 
     final editor = EditorClient(widget.dtd);
     ga.screen(PropertyEditorEvents.id);
-    unawaited(_editor = editor.initialized.then((_) => editor));
+    unawaited(
+      _editor = editor.initialized.then((_) {
+        _propertyEditorController = PropertyEditorController(editor);
+        return editor;
+      }),
+    );
+  }
+
+  @override
+  void dispose() {
+    _propertyEditorController?.dispose();
+    super.dispose();
   }
 
   @override
@@ -53,7 +64,10 @@ class _PropertyEditorSidebarPanelState
               snapshot.data,
             )) {
               (ConnectionState.done, final editor?) =>
-                _PropertyEditorConnectedPanel(editor),
+                _PropertyEditorConnectedPanel(
+                  editor,
+                  controller: _propertyEditorController!,
+                ),
               _ => const CenteredCircularProgressIndicator(),
             },
       ),
@@ -63,9 +77,10 @@ class _PropertyEditorSidebarPanelState
 
 /// The property editor panel shown once we know an editor is available.
 class _PropertyEditorConnectedPanel extends StatefulWidget {
-  const _PropertyEditorConnectedPanel(this.editor);
+  const _PropertyEditorConnectedPanel(this.editor, {required this.controller});
 
   final EditorClient editor;
+  final PropertyEditorController controller;
 
   @override
   State<_PropertyEditorConnectedPanel> createState() =>
@@ -96,8 +111,8 @@ class _PropertyEditorConnectedPanelState
       thumbVisibility: true,
       child: SingleChildScrollView(
         controller: scrollController,
-        child: const Padding(
-          padding: EdgeInsets.fromLTRB(
+        child: Padding(
+          padding: const EdgeInsets.fromLTRB(
             denseSpacing,
             defaultSpacing,
             defaultSpacing, // Additional right padding for scroll bar.
@@ -105,7 +120,7 @@ class _PropertyEditorConnectedPanelState
           ),
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
-            children: [PropertyEditorSidebar()],
+            children: [PropertyEditorView(controller: widget.controller)],
           ),
         ),
       ),