+++ packages/devtools_app/lib/src/standalone_ui/standalone_screen.dart
@@ -7,8 +7,8 @@ import 'package:flutter/material.dart';
 
 import '../shared/globals.dart';
 import '../shared/ui/common_widgets.dart';
+import 'ide_shared/property_editor/property_editor_panel.dart';
 import 'vs_code/flutter_panel.dart';
-import 'vs_code/property_editor_panel.dart';
 
 /// "Screens" that are intended for standalone use only, likely for embedding
 /// directly in an IDE.
@@ -50,7 +50,7 @@ enum StandaloneScreenType {
         builder: (context, data, _) {
           return _DtdConnectedScreen(
             dtd: data,
-            screenProvider: (dtd) => PropertyEditorSidebarPanel(dtd),
+            screenProvider: (dtd) => PropertyEditorPanel(dtd),
           );
         },
       ),